fix overflow in algorithm

+++ src/main/java/lesson3/task3/Solution.java
@@ -3,9 +3,9 @@ package lesson3.task3;
 public class Solution {
     public int solution(int[] A) {
         int n = A.length;
-        int missed = (n+1) * (n+2) / 2;
+        long missed = (n + 1L) * (n + 2) / 2;
         for (int i : A)
             missed -= i;
-        return missed;
+        return (int) missed;
     }
 }
